Add sync cache with db on apply

+++ app/models/db_tree_view.rb
@@ -9,8 +9,13 @@ class DbTreeView < ApplicationRecord
     transaction do
       data.each do |_index, node_data|
         new_node = find_or_initialize_by(id: node_data[:id])
+        # Don't save persisted disabled records
+        next if new_node.persisted? && new_node.disabled?
+
         new_node.assign_attributes(node_data.except(:id))
+        # Don't save new records without parent
         next if (new_node.parent.nil? || new_node.parent.disabled?) && !new_node.persisted?
+
         if new_node.disabled?
           new_node.destroy
         end
@@ -18,5 +23,9 @@ class DbTreeView < ApplicationRecord
       end
     end
     return errors.present? ? errors : true
+  end
+
+  def self.get_nodes(ids)
+    where(id: ids)
   end
 end
